+++ src/window_image_animation.h
@@ -1,0 +1,33 @@
+/**
+ * File:   window_image_animation.h
+ * Author: AWTK Develop Team
+ * Brief:  window_image_animation source
+ *
+ * Copyright (c) 2018 - 2021  Guangzhou ZHIYUAN Electronics Co.,Ltd.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * License file for more details.
+ *
+ */
+
+/**
+ * History:
+ * ================================================================
+ * 2021-05-19 Li XianJing <xianjimli@hotmail.com> created
+ *
+ */
+
+#ifndef WINDOW_IMAGE_ANIMATION_H
+#define WINDOW_IMAGE_ANIMATION_H
+
+#include "awtk.h"
+
+BEGIN_C_DECLS
+
+ret_t window_image_animation_open(void);
+
+END_C_DECLS
+
+#endif /*WINDOW_IMAGE_ANIMATION_H*/