+++ src/window_main.c
@@ -26,6 +26,7 @@
 #include "window_image.h"
 #include "window_font.h"
 #include "window_gif.h"
+#include "window_image_animation.h"
 #include "window_screen_saver.h"
 #include "window_check_button.h"
 #include "window_closable.h"
@@ -103,6 +104,11 @@ static ret_t on_open_font(void* ctx, event_t* e) {
 
 static ret_t on_open_gif(void* ctx, event_t* e) {
   window_gif_open();
+  return RET_OK;
+}
+
+static ret_t on_open_image_animation(void* ctx, event_t* e) {
+  window_image_animation_open();
   return RET_OK;
 }
 
@@ -196,6 +202,7 @@ ret_t window_main_open(void) {
   widget_child_on(win, "image", EVT_CLICK, on_open_image, win);
   widget_child_on(win, "font", EVT_CLICK, on_open_font, win);
   widget_child_on(win, "gif", EVT_CLICK, on_open_gif, win);
+  widget_child_on(win, "image_animation", EVT_CLICK, on_open_image_animation, win);
   widget_child_on(win, "screen_saver", EVT_CLICK, on_open_screen_saver, win);
   widget_child_on(win, "children_layout", EVT_CLICK, on_open_children_layout, win);
   widget_child_on(win, "self_layout", EVT_CLICK, on_open_self_layout, win);