+++ docs/changes.md
@@ -2,6 +2,7 @@
 
 2020/05/19
   * 增加screen saver测试。
+  * 增加image animation测试。
 
 2020/05/18
   * 增加image测试。